+++ packages/flutter/lib/src/widgets/_image_io.dart
@@ -1,0 +1,465 @@
+// Copyright 2015 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'dart:io';
+import 'dart:typed_data';
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter/painting.dart';
+import 'package:flutter/services.dart';
+
+import 'basic.dart';
+import 'framework.dart';
+import 'image.dart';
+import 'media_query.dart';
+
+export 'package:flutter/painting.dart' show
+  AssetImage,
+  ExactAssetImage,
+  FileImage,
+  FilterQuality,
+  ImageConfiguration,
+  ImageInfo,
+  ImageStream,
+  ImageProvider,
+  MemoryImage,
+  NetworkImage;
+
+/// A widget that displays an image.
+///
+/// Several constructors are provided for the various ways that an image can be
+/// specified:
+///
+///  * [new Image], for obtaining an image from an [ImageProvider].
+///  * [new Image.asset], for obtaining an image from an [AssetBundle]
+///    using a key.
+///  * [new Image.network], for obtaining an image from a URL.
+///  * [new Image.file], for obtaining an image from a [File].
+///  * [new Image.memory], for obtaining an image from a [Uint8List].
+///
+/// The following image formats are supported: {@macro flutter.dart:ui.imageFormats}
+///
+/// To automatically perform pixel-density-aware asset resolution, specify the
+/// image using an [AssetImage] and make sure that a [MaterialApp], [WidgetsApp],
+/// or [MediaQuery] widget exists above the [Image] widget in the widget tree.
+///
+/// The image is painted using [paintImage], which describes the meanings of the
+/// various fields on this class in more detail.
+///
+/// {@tool sample}
+/// The default constructor can be used with any [ImageProvider], such as a
+/// [NetworkImage], to display an image from the internet.
+///
+/// ![An image of an owl displayed by the image widget](https://flutter.github.io/assets-for-api-docs/assets/widgets/owl.jpg)
+///
+/// ```dart
+/// const Image(
+///   image: NetworkImage('https://flutter.github.io/assets-for-api-docs/assets/widgets/owl.jpg'),
+/// )
+/// ```
+/// {@end-tool}
+///
+/// {@tool sample}
+/// The [Image] Widget also provides several constructors to display different
+/// types of images for convenience. In this example, use the [Image.network]
+/// constructor to display an image from the internet.
+///
+/// ![An image of an owl displayed by the image widget using the shortcut constructor](https://flutter.github.io/assets-for-api-docs/assets/widgets/owl-2.jpg)
+///
+/// ```dart
+/// Image.network('https://flutter.github.io/assets-for-api-docs/assets/widgets/owl-2.jpg')
+/// ```
+/// {@end-tool}
+///
+/// See also:
+///
+///  * [Icon], which shows an image from a font.
+///  * [new Ink.image], which is the preferred way to show an image in a
+///    material application (especially if the image is in a [Material] and will
+///    have an [InkWell] on top of it).
+///  * [Image](dart-ui/Image-class.html), the class in the [dart:ui] library.
+///
+class Image extends ImageBase {
+  /// Creates a widget that displays an image.
+  ///
+  /// To show an image from the network or from an asset bundle, consider using
+  /// [new Image.network] and [new Image.asset] respectively.
+  ///
+  /// The [image], [alignment], [repeat], and [matchTextDirection] arguments
+  /// must not be null.
+  ///
+  /// Either the [width] and [height] arguments should be specified, or the
+  /// widget should be placed in a context that sets tight layout constraints.
+  /// Otherwise, the image dimensions will change as the image is loaded, which
+  /// will result in ugly layout changes.
+  ///
+  /// Use [filterQuality] to change the quality when scaling an image.
+  /// Use the [FilterQuality.low] quality setting to scale the image,
+  /// which corresponds to bilinear interpolation, rather than the default
+  /// [FilterQuality.none] which corresponds to nearest-neighbor.
+  ///
+  /// If [excludeFromSemantics] is true, then [semanticLabel] will be ignored.
+  const Image({
+    Key key,
+    @required this.image,
+    this.frameBuilder,
+    this.loadingBuilder,
+    this.semanticLabel,
+    this.excludeFromSemantics = false,
+    this.width,
+    this.height,
+    this.color,
+    this.colorBlendMode,
+    this.fit,
+    this.alignment = Alignment.center,
+    this.repeat = ImageRepeat.noRepeat,
+    this.centerSlice,
+    this.matchTextDirection = false,
+    this.gaplessPlayback = false,
+    this.filterQuality = FilterQuality.low,
+  }) : assert(image != null),
+       assert(alignment != null),
+       assert(repeat != null),
+       assert(filterQuality != null),
+       assert(matchTextDirection != null),
+       super(key: key);
+
+  /// Creates a widget that displays an [ImageStream] obtained from the network.
+  ///
+  /// The [src], [scale], and [repeat] arguments must not be null.
+  ///
+  /// Either the [width] and [height] arguments should be specified, or the
+  /// widget should be placed in a context that sets tight layout constraints.
+  /// Otherwise, the image dimensions will change as the image is loaded, which
+  /// will result in ugly layout changes.
+  ///
+  /// All network images are cached regardless of HTTP headers.
+  ///
+  /// An optional [headers] argument can be used to send custom HTTP headers
+  /// with the image request.
+  ///
+  /// Use [filterQuality] to change the quality when scaling an image.
+  /// Use the [FilterQuality.low] quality setting to scale the image,
+  /// which corresponds to bilinear interpolation, rather than the default
+  /// [FilterQuality.none] which corresponds to nearest-neighbor.
+  ///
+  /// If [excludeFromSemantics] is true, then [semanticLabel] will be ignored.
+  Image.network(
+    String src, {
+    Key key,
+    double scale = 1.0,
+    this.frameBuilder,
+    this.loadingBuilder,
+    this.semanticLabel,
+    this.excludeFromSemantics = false,
+    this.width,
+    this.height,
+    this.color,
+    this.colorBlendMode,
+    this.fit,
+    this.alignment = Alignment.center,
+    this.repeat = ImageRepeat.noRepeat,
+    this.centerSlice,
+    this.matchTextDirection = false,
+    this.gaplessPlayback = false,
+    this.filterQuality = FilterQuality.low,
+    Map<String, String> headers,
+  }) : image = NetworkImage(src, scale: scale, headers: headers),
+       assert(alignment != null),
+       assert(repeat != null),
+       assert(matchTextDirection != null),
+       super(key: key);
+
+  /// Creates a widget that displays an [ImageStream] obtained from a [File].
+  ///
+  /// This class is not supported in Flutter for web applications.
+  ///
+  /// The [file], [scale], and [repeat] arguments must not be null.
+  ///
+  /// Either the [width] and [height] arguments should be specified, or the
+  /// widget should be placed in a context that sets tight layout constraints.
+  /// Otherwise, the image dimensions will change as the image is loaded, which
+  /// will result in ugly layout changes.
+  ///
+  /// On Android, this may require the
+  /// `android.permission.READ_EXTERNAL_STORAGE` permission.
+  ///
+  /// Use [filterQuality] to change the quality when scaling an image.
+  /// Use the [FilterQuality.low] quality setting to scale the image,
+  /// which corresponds to bilinear interpolation, rather than the default
+  /// [FilterQuality.none] which corresponds to nearest-neighbor.
+  ///
+  /// If [excludeFromSemantics] is true, then [semanticLabel] will be ignored.
+  Image.file(
+    File file, {
+    Key key,
+    double scale = 1.0,
+    this.frameBuilder,
+    this.semanticLabel,
+    this.excludeFromSemantics = false,
+    this.width,
+    this.height,
+    this.color,
+    this.colorBlendMode,
+    this.fit,
+    this.alignment = Alignment.center,
+    this.repeat = ImageRepeat.noRepeat,
+    this.centerSlice,
+    this.matchTextDirection = false,
+    this.gaplessPlayback = false,
+    this.filterQuality = FilterQuality.low,
+  }) : image = FileImage(file, scale: scale),
+       loadingBuilder = null,
+       assert(alignment != null),
+       assert(repeat != null),
+       assert(filterQuality != null),
+       assert(matchTextDirection != null),
+       super(key: key);
+
+
+  // TODO(ianh): Implement the following (see ../services/image_resolution.dart):
+  //
+  // * If [width] and [height] are both specified, and [scale] is not, then
+  //   size-aware asset resolution will be attempted also, with the given
+  //   dimensions interpreted as logical pixels.
+  //
+  // * If the images have platform, locale, or directionality variants, the
+  //   current platform, locale, and directionality are taken into account
+  //   during asset resolution as well.
+  /// Creates a widget that displays an [ImageStream] obtained from an asset
+  /// bundle. The key for the image is given by the `name` argument.
+  ///
+  /// The `package` argument must be non-null when displaying an image from a
+  /// package and null otherwise. See the `Assets in packages` section for
+  /// details.
+  ///
+  /// If the `bundle` argument is omitted or null, then the
+  /// [DefaultAssetBundle] will be used.
+  ///
+  /// By default, the pixel-density-aware asset resolution will be attempted. In
+  /// addition:
+  ///
+  /// * If the `scale` argument is provided and is not null, then the exact
+  /// asset specified will be used. To display an image variant with a specific
+  /// density, the exact path must be provided (e.g. `images/2x/cat.png`).
+  ///
+  /// If [excludeFromSemantics] is true, then [semanticLabel] will be ignored.
+  ///
+  /// The [name] and [repeat] arguments must not be null.
+  ///
+  /// Either the [width] and [height] arguments should be specified, or the
+  /// widget should be placed in a context that sets tight layout constraints.
+  /// Otherwise, the image dimensions will change as the image is loaded, which
+  /// will result in ugly layout changes.
+  ///
+  /// Use [filterQuality] to change the quality when scaling an image.
+  /// Use the [FilterQuality.low] quality setting to scale the image,
+  /// which corresponds to bilinear interpolation, rather than the default
+  /// [FilterQuality.none] which corresponds to nearest-neighbor.
+  ///
+  /// {@tool sample}
+  ///
+  /// Suppose that the project's `pubspec.yaml` file contains the following:
+  ///
+  /// ```yaml
+  /// flutter:
+  ///   assets:
+  ///     - images/cat.png
+  ///     - images/2x/cat.png
+  ///     - images/3.5x/cat.png
+  /// ```
+  /// {@end-tool}
+  ///
+  /// On a screen with a device pixel ratio of 2.0, the following widget would
+  /// render the `images/2x/cat.png` file:
+  ///
+  /// ```dart
+  /// Image.asset('images/cat.png')
+  /// ```
+  ///
+  /// This corresponds to the file that is in the project's `images/2x/`
+  /// directory with the name `cat.png` (the paths are relative to the
+  /// `pubspec.yaml` file).
+  ///
+  /// On a device with a 4.0 device pixel ratio, the `images/3.5x/cat.png` asset
+  /// would be used. On a device with a 1.0 device pixel ratio, the
+  /// `images/cat.png` resource would be used.
+  ///
+  /// The `images/cat.png` image can be omitted from disk (though it must still
+  /// be present in the manifest). If it is omitted, then on a device with a 1.0
+  /// device pixel ratio, the `images/2x/cat.png` image would be used instead.
+  ///
+  ///
+  /// ## Assets in packages
+  ///
+  /// To create the widget with an asset from a package, the [package] argument
+  /// must be provided. For instance, suppose a package called `my_icons` has
+  /// `icons/heart.png` .
+  ///
+  /// {@tool sample}
+  /// Then to display the image, use:
+  ///
+  /// ```dart
+  /// Image.asset('icons/heart.png', package: 'my_icons')
+  /// ```
+  /// {@end-tool}
+  ///
+  /// Assets used by the package itself should also be displayed using the
+  /// [package] argument as above.
+  ///
+  /// If the desired asset is specified in the `pubspec.yaml` of the package, it
+  /// is bundled automatically with the app. In particular, assets used by the
+  /// package itself must be specified in its `pubspec.yaml`.
+  ///
+  /// A package can also choose to have assets in its 'lib/' folder that are not
+  /// specified in its `pubspec.yaml`. In this case for those images to be
+  /// bundled, the app has to specify which ones to include. For instance a
+  /// package named `fancy_backgrounds` could have:
+  ///
+  /// ```
+  /// lib/backgrounds/background1.png
+  /// lib/backgrounds/background2.png
+  /// lib/backgrounds/background3.png
+  /// ```
+  ///
+  /// To include, say the first image, the `pubspec.yaml` of the app should
+  /// specify it in the assets section:
+  ///
+  /// ```yaml
+  ///   assets:
+  ///     - packages/fancy_backgrounds/backgrounds/background1.png
+  /// ```
+  ///
+  /// The `lib/` is implied, so it should not be included in the asset path.
+  ///
+  ///
+  /// See also:
+  ///
+  ///  * [AssetImage], which is used to implement the behavior when the scale is
+  ///    omitted.
+  ///  * [ExactAssetImage], which is used to implement the behavior when the
+  ///    scale is present.
+  ///  * <https://flutter.dev/assets-and-images/>, an introduction to assets in
+  ///    Flutter.
+  Image.asset(
+    String name, {
+    Key key,
+    AssetBundle bundle,
+    this.frameBuilder,
+    this.semanticLabel,
+    this.excludeFromSemantics = false,
+    double scale,
+    this.width,
+    this.height,
+    this.color,
+    this.colorBlendMode,
+    this.fit,
+    this.alignment = Alignment.center,
+    this.repeat = ImageRepeat.noRepeat,
+    this.centerSlice,
+    this.matchTextDirection = false,
+    this.gaplessPlayback = false,
+    String package,
+    this.filterQuality = FilterQuality.low,
+  }) : image = scale != null
+         ? ExactAssetImage(name, bundle: bundle, scale: scale, package: package)
+         : AssetImage(name, bundle: bundle, package: package),
+       loadingBuilder = null,
+       assert(alignment != null),
+       assert(repeat != null),
+       assert(matchTextDirection != null),
+       super(key: key);
+
+  /// Creates a widget that displays an [ImageStream] obtained from a [Uint8List].
+  ///
+  /// The [bytes], [scale], and [repeat] arguments must not be null.
+  ///
+  /// This only accepts compressed image formats (e.g. PNG). Uncompressed
+  /// formats like rawRgba (the default format of [ui.Image.toByteData]) will
+  /// lead to exceptions.
+  ///
+  /// Either the [width] and [height] arguments should be specified, or the
+  /// widget should be placed in a context that sets tight layout constraints.
+  /// Otherwise, the image dimensions will change as the image is loaded, which
+  /// will result in ugly layout changes.
+  ///
+  /// Use [filterQuality] to change the quality when scaling an image.
+  /// Use the [FilterQuality.low] quality setting to scale the image,
+  /// which corresponds to bilinear interpolation, rather than the default
+  /// [FilterQuality.none] which corresponds to nearest-neighbor.
+  ///
+  /// If [excludeFromSemantics] is true, then [semanticLabel] will be ignored.
+  Image.memory(
+    Uint8List bytes, {
+    Key key,
+    double scale = 1.0,
+    this.frameBuilder,
+    this.semanticLabel,
+    this.excludeFromSemantics = false,
+    this.width,
+    this.height,
+    this.color,
+    this.colorBlendMode,
+    this.fit,
+    this.alignment = Alignment.center,
+    this.repeat = ImageRepeat.noRepeat,
+    this.centerSlice,
+    this.matchTextDirection = false,
+    this.gaplessPlayback = false,
+    this.filterQuality = FilterQuality.low,
+  }) : image = MemoryImage(bytes, scale: scale),
+       loadingBuilder = null,
+       assert(alignment != null),
+       assert(repeat != null),
+       assert(matchTextDirection != null),
+       super(key: key);
+
+    @override
+    final ImageProvider image;
+
+    @override
+    final ImageFrameBuilder frameBuilder;
+
+    @override
+    final String semanticLabel;
+
+    @override
+    final bool excludeFromSemantics;
+
+    @override
+    final double width;
+
+    @override
+    final double height;
+
+    @override
+    final Color color;
+
+    @override
+    final BlendMode colorBlendMode;
+
+    @override
+    final BoxFit fit;
+
+    @override
+    final AlignmentGeometry alignment;
+
+    @override
+    final ImageRepeat repeat;
+
+    @override
+    final Rect centerSlice;
+
+    @override
+    final bool matchTextDirection;
+
+    @override
+    final bool gaplessPlayback;
+
+    @override
+    final FilterQuality filterQuality;
+
+    @override
+    final ImageLoadingBuilder loadingBuilder;
+}
